Feature/backend user

+++ backend/src/main/java/com/nocountry/backend/dto/video/VideoOutput.java
@@ -3,7 +3,7 @@ package com.nocountry.backend.dto.video;
 public record VideoOutput(
         String fileName,
         String path,
-        Long durationSeconds,
+        Integer durationSeconds,
         Long videoSizeBytes
 ) {
 }
